fix: import error due to missing prefix, move util funcs into utils.py
